Fixed Automatic Naming Bug
Fixed a bug where if a new star had the same filename as an old star there would be an inifite loop in naming. The issue was caused because the star names aren't given file extensions but the program was looking for an extension when naming.

The default naming for a star was changed as well to always use one default and then use the unlimited naming section to renumber the defualt as needed.