Bisection is working now
the code runs but the bisect function reuqires that Lum_error(a) and Lum_error(b) have opposite sign in order to find the root. apparently they dont. it also requires input for the  radius, surfae temp and the energy generation. I havent figured out a way to add these as arguments of Lum_error bc calling Lum_error in optimize.bisect requires the functon to have one input where is plugs in a and b.